Создал модули в паке JS
